requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiofiles==23.2.1
python-mimeparse==1.6.0
email-validator==2.1.0
jinja2==3.1.2
aiosmtplib==3.0.1
cryptography==41.0.7
itsdangerous==2.1.2

# Production dependencies
gunicorn==21.2.0
redis==5.0.1
slowapi==0.1.9
sentry-sdk[fastapi]==1.38.0
prometheus-fastapi-instrumentator==6.1.0

# AWS SDK (if using S3)
boto3==1.34.0
botocore==1.34.0

# Monitoring and logging
structlog==23.2.0
python-json-logger==2.0.7
